#18212745 #18210564 fix Guestbook reply link and admin delete
git-svn-id: http://xe-textyle.googlecode.com/svn/sandbox/textyle@27 bebf4043-ad85-6e1f-322f-7830e3d0fd93

+++ textyle.controller.php
@@ -436,11 +436,11 @@
          * @brief Guestbook item delete
 		 **/
 		function procTextyleGuestbookItemDelete(){
-            $oTextyleModel = &getModel('textyle');
 			$textyle_guestbook_srl = Context::get('textyle_guestbook_srl');
 			if(!$textyle_guestbook_srl) return new Object(-1,'msg_invalid_request');
-   
-			if(!$_SESSION['own_textyle_guestbook'][$val->textyle_guestbook_srl]) return new Object(-1,'msg_not_permitted');
+
+			$logged_info = Context::get('logged_info');  
+			if(!($logged_info->is_site_admin || $_SESSION['own_textyle_guestbook'][$val->textyle_guestbook_srl])) return new Object(-1,'msg_not_permitted');
 			$output = $this->deleteGuestbookItem($textyle_guestbook_srl);
 			return $output;
 		}
